Clean code, searches delete on long press, old searches display and on click search again.
