+++ SprintPokemonChallenge/app/src/main/java/com/example/sprintpokemonchallenge/model/PokemonIndex.kt
@@ -2,10 +2,6 @@ package com.example.sprintpokemonchallenge.model
 
 import com.google.gson.annotations.SerializedName
 
-data class PokemonIndex(val count: Int, val results: List<Pokemon>)
-
-data class Pokemon(val name: String, val url: String)
-
 data class PokemonDetail(val abilities: List<Ability>, val id: Int, val name: String, val sprites: Sprites, val types: List<Types>)
 //Next five data class's help flesh out the above class, See parameters
 data class Ability(val ability: AbilityDetails)
