+++ SprintPokemonChallenge/app/src/main/java/com/example/sprintpokemonchallenge/retrofit/RetrofitInstance.kt
@@ -2,7 +2,6 @@ package com.example.sprintpokemonchallenge.retrofit
 
 import com.example.sprintpokemonchallenge.api.PokemonApi
 import com.example.sprintpokemonchallenge.model.PokemonDetail
-import com.example.sprintpokemonchallenge.model.PokemonIndex
 import com.google.gson.GsonBuilder
 import retrofit2.Call
 import retrofit2.Retrofit
@@ -10,20 +9,6 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 object RetrofitInstance{
     private const val BASE_URL = "https://pokeapi.co/api/v2/pokemon/"
-
-    fun getPokemonIndex(): Call<PokemonIndex>{
-        val gson = GsonBuilder()
-            .setLenient()
-            .create()
-
-        val retrofit = Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create(gson))
-            .build()
-
-        return retrofit.create(PokemonApi::class.java).getPokemonIndex()
-
-    }
 
     fun getPokemon(pokemonNameOrId: String): Call<PokemonDetail>{
         val gson = GsonBuilder()
