+++ SprintPokemonChallenge/app/src/main/java/com/example/sprintpokemonchallenge/ui/ListOfPokemon.kt
@@ -2,12 +2,7 @@ package com.example.sprintpokemonchallenge.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.example.sprintpokemonchallenge.R
-import com.example.sprintpokemonchallenge.recyclerAdapter.RecyclerAdapter
-import com.example.sprintpokemonchallenge.ui.MainActivity.Companion.pokemonList
-import kotlinx.android.synthetic.main.activity_list_of_pokemon.*
 
 class ListOfPokemon : AppCompatActivity() {
 
@@ -15,10 +10,10 @@ class ListOfPokemon : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_list_of_pokemon)
 
-        recycler.setHasFixedSize(true)
-        val manager = LinearLayoutManager(this, RecyclerView.VERTICAL, false)
-        val adapter = RecyclerAdapter(pokemonList)
-        recycler.layoutManager = manager
-        recycler.adapter = adapter
+//        recycler.setHasFixedSize(true)
+//        val manager = LinearLayoutManager(this, RecyclerView.VERTICAL, false)
+//        val adapter = RecyclerAdapter(pokemonList)
+//        recycler.layoutManager = manager
+//        recycler.adapter = adapter
     }
 }
