+++ SprintPokemonChallenge/app/src/main/java/com/example/sprintpokemonchallenge/ui/MainActivity.kt
@@ -3,34 +3,26 @@ package com.example.sprintpokemonchallenge.ui
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.example.sprintpokemonchallenge.R
-import com.example.sprintpokemonchallenge.model.Pokemon
 import com.example.sprintpokemonchallenge.model.PokemonDetail
-import com.example.sprintpokemonchallenge.model.PokemonIndex
+import com.example.sprintpokemonchallenge.recyclerAdapter.RecyclerAdapter
 import com.example.sprintpokemonchallenge.retrofit.RetrofitInstance
 import kotlinx.android.synthetic.main.activity_main.*
+import kotlinx.android.synthetic.main.activity_main.recycler
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class MainActivity : AppCompatActivity(), Callback<PokemonIndex> {
+class MainActivity : AppCompatActivity() {
 
     companion object {
-        var pokemonList: List<Pokemon>? = listOf()
         lateinit var pokemonInfo: PokemonDetail
+        var searchedPokemon: ArrayList<PokemonDetail> = arrayListOf()
     }
 
-    override fun onFailure(call: Call<PokemonIndex>, t: Throwable) {
-        Log.i("IsWorking", "No")
-    }
-
-    override fun onResponse(call: Call<PokemonIndex>, response: Response<PokemonIndex>) {
-        val intent = Intent(this, ListOfPokemon::class.java)
-        pokemonList = response.body()?.results
-        startActivity(intent)
-    }
-
+    
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
@@ -44,6 +36,23 @@ class MainActivity : AppCompatActivity(), Callback<PokemonIndex> {
                     if (response.body() != null){
                         pokemonInfo = response.body() as PokemonDetail
 
+                        var addPokemonToList = true
+                        if (searchedPokemon.size < 1){
+                            searchedPokemon.add(pokemonInfo)
+                            addPokemonToList = false
+                        }
+                        else {
+                            for (i in searchedPokemon.indices){
+                                if(searchedPokemon[i] == pokemonInfo){
+                                    addPokemonToList = false
+                                    break
+                                }
+                            }
+                        }
+                        if (addPokemonToList){
+                            searchedPokemon.add(pokemonInfo)
+                        }
+
                         val intent = Intent(this@MainActivity, PokemonDetailsActivity::class.java)
                         startActivity(intent)
                     }
@@ -52,8 +61,16 @@ class MainActivity : AppCompatActivity(), Callback<PokemonIndex> {
 
         }
 
-        btn_list.setOnClickListener {
-            RetrofitInstance.getPokemonIndex().enqueue(this)
-        }
+    }
+
+    override fun onResume() {
+        recycler.setHasFixedSize(true)
+        val manager = LinearLayoutManager(this, RecyclerView.VERTICAL, false)
+        recycler.layoutManager = manager
+        val adapter = RecyclerAdapter(searchedPokemon)
+        recycler.adapter = adapter
+        super.onResume()
+
+//        edit_txt.setText()
     }
 }
