+++ SprintPokemonChallenge/app/src/main/res/layout/activity_list_of_pokemon.xml
@@ -7,7 +7,7 @@
     tools:context=".ui.ListOfPokemon">
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler"
+        android:id="@+id/recycler2"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:listitem="@layout/recycler_item"
